feat: the MFA toggle now sends a post request to the backend when the profile page is unloaded
