+++ app/Views/home.latte
@@ -14,8 +14,8 @@
     'title' => 'Quick Actions',
     'template' => 'actions.latte','args' => [
         'actions' => [
-            ['icon' => 'plus-circle', 'color' => 'success', 'label' => 'Create a vault', 'target' => '#create-vault'],
-            ['icon' => 'sync-alt', 'color' => 'primary', 'label' => 'Generate Vault'],
+            ['icon' => 'plus-circle', 'color' => 'success', 'label' => 'Create vault', 'target' => '#create-vault'],
+            ['icon' => 'sync-alt', 'color' => 'primary', 'label' => 'Generate vault'],
             ['icon' => 'share-alt', 'color' => 'info', 'label' => 'Share vault'],
         ]
     ]
