+++ app/Views/layouts/components/profile/multifactor_toggle.latte
@@ -1,7 +1,7 @@
 <div class="d-flex flex-column px-0">
     <div class="d-flex form-check form-switch justify-content-between align-items-center px-0">
         <label class="form-check-label" for="mfaToggle">Enable Multifactor Authentication</label>
-        <input class="form-check-input" type="checkbox" id="mfaToggle">
+        <input id="mfaToggle" class="form-check-input" type="checkbox" data-mfa checked>
     </div>
     <small class="text-muted mt-1">
         Add an extra layer of security to your account by requiring a second form of verification during login.
