+++ public/javascripts/main.js
@@ -1,6 +1,7 @@
 import listen from './navbar.js'
 import Application from './app.js';
-import {togglePassword, focusOnModal} from "./vault.js";
+import {togglePassword} from "./vault.js";
+import saveMFA from "./profile.js";
 
 activateListeners();
 new Application().initialize();
@@ -8,5 +9,25 @@ new Application().initialize();
 function activateListeners() {
     listen();
     togglePassword();
-    //focusOnManage()
+    focusOnModal();
+    saveMFA();
+}
+
+
+function focusOnModal() {
+    document.querySelectorAll('[data-trigger]').forEach(btn => {
+        btn.addEventListener('click', () => {
+            const modalSelector = btn.getAttribute('data-bs-target');
+            const modal = document.querySelector(modalSelector);
+
+            if (!modal) return;
+
+            const instance = bootstrap.Modal.getOrCreateInstance(modal);
+            instance.show();
+
+            modal.addEventListener('shown.bs.modal', () => {
+                modal.querySelector('input')?.focus();
+            }, {once: true});
+        });
+    });
 }
