+++ public/stylesheets/components/headline.css
@@ -1,4 +1,4 @@
-h1, h2, h3, h4, h5 {
+h1, h2, h3, h4, h5, h6 {
     font-family: var(--headline-font), sans-serif;
     margin: 0;
 }
